Implemented writeNFiles All operations working properly File saving (both read and miss) working properly minor refactoring minor bugfixing

+++ headers/serv_manager.h
@@ -22,7 +22,7 @@ sc_cache* server_cache;      // the actual cache structure used by the server
 size_t server_byte_size;      // indicates the capacity of the server in terms of bytes of memory
 size_t server_file_size;      // indicates the capacity of the server in terms of files in memory
 
-unsigned thread_pool_cap;      // indicates the maximum number of worker-threads the server can manage at the same time
+unsigned long thread_pool_cap;      // indicates the maximum number of worker-threads the server can manage at the same time
 char* sock_addr;       // will hold the server's main socket address
 conc_queue* requests_queue;        // the queue worker-threads will use to dispatch client's requests TODO cambiare tipo
 pthread_t* worker_threads_arr;      // array of worker threads
